+++ kubernetes/scripts/api/api.sh
@@ -27,16 +27,16 @@ case $CMD in
   "ls")
     case "$ANALYSIS_ID" in
       "p"|"portfolio"|"portfolios")
-        $CURL -H "$CAH" -X GET "${API_URL}/v1/portfolios/" | jq -r '.[] | "\(.id) - \(.name)\r\t\t\t\tgroups: \(.groups | join (","))"' | sort -n
+        curlf -X GET "${API_URL}/v1/portfolios/" | jq -r '.[] | "\(.id) - \(.name)\r\t\t\t\tgroups: \(.groups | join (","))"' | sort -n
       ;;
       "m"|"model"|"models")
-        $CURL -H "$CAH" -X GET "${API_URL}/v1/models/" | jq -r '.[] | "\(.id) - \(.supplier_id)-\(.model_id)-\(.version_id)\r\t\t\t\tgroups: \(.groups | join (","))"' | sort -n
+        curlf -X GET "${API_URL}/v1/models/" | jq -r '.[] | "\(.id) - \(.supplier_id)-\(.model_id)-\(.version_id)\r\t\t\t\tgroups: \(.groups | join (","))"' | sort -n
       ;;
       "df"|"data-files"|"datafiles")
-        $CURL -H "$CAH" -X GET "${API_URL}/v1/data_files/" | jq -r '.[] | "\(.id) - \(.filename)\r\t\t\t\tgroups: \(.groups | join (","))"'| sort -n
+        curlf -X GET "${API_URL}/v1/data_files/" | jq -r '.[] | "\(.id) - \(.filename)\r\t\t\t\tgroups: \(.groups | join (","))"'| sort -n
       ;;
       *)
-        $CURL -H "$CAH" -X GET "${API_URL}/v1/analyses/" | jq -r '.[] | "\(.id) - \(.status) - \(.name)\r\t\t\t\t\t\tportfolio: \(.portfolio)\tmodel: \(.model)\tgroups: \(.groups | join (","))"' | sort -n
+        curlf -X GET "${API_URL}/v1/analyses/" | jq -r '.[] | "\(.id) - \(.status) - \(.name)\r\t\t\t\t\t\tportfolio: \(.portfolio)\tmodel: \(.model)\tgroups: \(.groups | join (","))"' | sort -n
       ;;
     esac
   ;;
@@ -60,11 +60,11 @@ case $CMD in
         PARAMS+=("\"supplier_id\": \"jxn\"")
         PARAMS+=("\"model_id\": \"jxn_catastrophy\"")
         PARAMS+=("\"version_id\": \"${NAME}\"")
-        $CURL -H "$CAH" -X POST "${API_URL}/v1/models/" -H  "Content-Type: application/json" -d "{$(join , "${PARAMS[@]}")}" | jq .
+        curlf -X POST "${API_URL}/v1/models/" -H  "Content-Type: application/json" -d "{$(join , "${PARAMS[@]}")}" | jq .
       ;;
       "portfolio")
         PARAMS+=("\"name\": \"${NAME}\"")
-        $CURL -H "$CAH" -X POST "${API_URL}/v1/portfolios/" -H  "Content-Type: application/json" -d "{$(join , "${PARAMS[@]}")}" | jq .
+        curlf -X POST "${API_URL}/v1/portfolios/" -H  "Content-Type: application/json" -d "{$(join , "${PARAMS[@]}")}" | jq .
       ;;
       "df")
         PARAMS+=("\"file_description\": \"${NAME}\"")
@@ -75,7 +75,7 @@ case $CMD in
           exit 1
         fi
         echo "Data file created with id: $DFID"
-        echo "" | $CURL -H "$CAH" -X POST "${API_URL}/v1/data_files/${DFID}/content/" -H  "Content-Type: multipart/form-data" -F "file=@-;filename=${NAME};type=application/json" | jq .
+        echo "" | curlf -X POST "${API_URL}/v1/data_files/${DFID}/content/" -H  "Content-Type: multipart/form-data" -F "file=@-;filename=${NAME};type=application/json" | jq .
       ;;
     esac
   ;;
@@ -92,7 +92,10 @@ case $CMD in
 
       case "$2" in
         "df")
-          $CURL -H "$CAH" -X DELETE -s -o /dev/null -w "%{http_code}" "${API_URL}/v1/data_files/${id}/"
+          curlf -X DELETE -s -o /dev/null -w "%{http_code}" "${API_URL}/v1/data_files/${id}/"
+        ;;
+        "a")
+          curlf -X DELETE -s -o /dev/null -w "%{http_code}" "${API_URL}/v1/analyses/${id}/"
         ;;
       esac
 
@@ -105,21 +108,21 @@ case $CMD in
       usage
     fi
 
-    $CURL -H "$CAH" -X POST "${API_URL}/v1/analyses/${ANALYSIS_ID}/cancel/" | jq -r '.status[0]'
+    curlf -X POST "${API_URL}/v1/analyses/${ANALYSIS_ID}/cancel/" | jq -r '.status'
   ;;
-  "generate")
+  "generate"|"input")
     if [ -z "$ANALYSIS_ID" ]; then
       usage
     fi
 
-    $CURL -H "$CAH" -X POST "${API_URL}/v1/analyses/${ANALYSIS_ID}/generate_inputs/" | jq -r '.status'
+    curlf -X POST "${API_URL}/v1/analyses/${ANALYSIS_ID}/generate_inputs/" | jq -r '.status'
   ;;
-  "execute")
+  "execute"|"loss")
     if [ -z "$ANALYSIS_ID" ]; then
       usage
     fi
 
-    $CURL -H "$CAH" -X POST "${API_URL}/v1/analyses/${ANALYSIS_ID}/run/" | jq -r '.status'
+    curlf -X POST "${API_URL}/v1/analyses/${ANALYSIS_ID}/run/" | jq -r '.status'
   ;;
   "run")
 
@@ -127,23 +130,45 @@ case $CMD in
       usage
     fi
 
-    $0 generate $ANALYSIS_ID
+    if [ "${#@}" == 2 ]; then
 
-    while :; do
-      STATUS=$($CURL -H "$CAH" -X GET "${API_URL}/v1/analyses/${ANALYSIS_ID}/" | jq -r '.status')
-      echo $STATUS
+      echo "$ANALYSIS_ID: $($0 input $ANALYSIS_ID) - input start"
 
-      if [ $STATUS == "READY" ]; then
-        $0 execute $ANALYSIS_ID
-      elif [ $STATUS == "RUN_COMPLETED" ] || [ $STATUS == "INPUTS_GENERATION_ERROR" ] || [ $STATUS == "RUN_ERROR" ]; then
-        break
-      fi
+      while :; do
+        STATUS=$($CURL -H "$CAH" -X GET "${API_URL}/v1/analyses/${ANALYSIS_ID}/" | jq -r '.status')
+        echo "$ANALYSIS_ID: $STATUS"
 
-      sleep 2
-    done
+        if [ $STATUS == "READY" ]; then
+          echo "$ANALYSIS_ID: $($0 loss $ANALYSIS_ID) - loss start"
+        elif [ $STATUS == "RUN_COMPLETED" ] || [ $STATUS == "INPUTS_GENERATION_ERROR" ] || [ $STATUS == "RUN_ERROR" ]; then
+          break
+        fi
+
+        sleep 2
+      done
+
+    else
+
+      PS=()
+      function cleanup()
+      {
+          echo "${PS[@]}"
+          kill "${PS[@]}"
+          exit
+      }
+
+      trap cleanup SIGINT
+
+      for i in "${@:2}"; do
+        $0 run $i &
+        PS+=($!)
+      done
+
+      wait
+    fi
   ;;
   "queue-status"|"qs")
-    $CURL -H "$CAH" -X GET "${API_URL}/v1/queue-status/" | jq .
+    curlf -X GET "${API_URL}/v1/queue-status/" | jq .
   ;;
   "token")
     echo $ACCESS_TOKEN