+++ kubernetes/worker-controller/src/worker_controller.py
@@ -40,18 +40,20 @@ def parse_args():
     :return:
     """
     parser = argparse.ArgumentParser('Oasis example model worker controller')
-    parser.add_argument('--api-host', help='The sever API hostname.', default=getenv('OASIS_API_HOST') or 'localhost')
-    parser.add_argument('--api-port', help='The server API portnumber.', default=getenv('OASIS_API_PORT') or 8000)
-    parser.add_argument('--secure', help='Flag if https and wss should be used.', default=bool(getenv('OASIS_API_SECURE')), action='store_true')
-    parser.add_argument('--username', help='The username of the worker controller user.', default=getenv('OASIS_USERNAME') or 'admin')
-    parser.add_argument('--password', help='The password of the worker controller user.', default=getenv('OASIS_PASSWORD') or 'password')
+    parser.add_argument('--api-host', help='The sever API hostname', default=getenv('OASIS_API_HOST') or 'localhost')
+    parser.add_argument('--api-port', help='The server API portnumber', default=getenv('OASIS_API_PORT') or 8000)
+    parser.add_argument('--secure', help='Flag if https and wss should be used', default=bool(getenv('OASIS_API_SECURE')), action='store_true')
+    parser.add_argument('--username', help='The username of the worker controller user', default=getenv('OASIS_USERNAME') or 'admin')
+    parser.add_argument('--password', help='The password of the worker controller user', default=getenv('OASIS_PASSWORD') or 'password')
+    parser.add_argument('--limit', help='Hard limit for the total number of workers created', default=getenv('OASIS_TOTAL_WORKER_LIMIT'))
+    parser.add_argument('--prioritized-models-limit', help='When prioritized runs are used - create workers for the models with the highest priority', default=getenv('OASIS_PRIORITIZED_MODELS_LIMIT'))
     parser.add_argument('--cluster', help='Type of kubernetes cluster to connect to, either "local" (~/.kube/config)\
-     or "in" to connect to the cluster the pod exists in.', default=getenv('CLUSTER') or 'in')
+     or "in" to connect to the cluster the pod exists in', default=getenv('CLUSTER') or 'in')
 
     args = parser.parse_args()
 
     for key in args.__dict__:
-        if args.__dict__[key] is None:
+        if args.__dict__[key] is None and key not in ['limit', 'prioritized_models_limit']:
             raise Exception(f'Missing value for {key}')
 
     if args.cluster != 'in' and args.cluster != 'local':
@@ -80,7 +82,7 @@ def main():
     event_loop.run_until_complete(cluster_client.load_config(args.cluster))
 
     # Create the autoscaler to bind everything together
-    autoscaler = AutoScaler(deployments, cluster_client, oasis_client)
+    autoscaler = AutoScaler(deployments, cluster_client, oasis_client, args.prioritized_models_limit, args.limit)
 
     # Create the deployment watcher and load all available deployments
     deployments_watcher = DeploymentWatcher(deployments)