+++ src/server/oasisapi/celery_app.py
@@ -2,7 +2,6 @@ from __future__ import absolute_import
 
 import os
 
-# from src.server.oasisapi.celeryimport.celeryimport import Celery
 from celery import Celery
 from django.conf import settings
 
@@ -11,3 +10,7 @@ os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'src.server.oasisapi.settings')
 celery_app = Celery('oasisapi')
 celery_app.config_from_object('django.conf:settings')
 celery_app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
+celery_app.conf.broker_transport_options = {
+    'priority_steps': list(range(9)),
+    'queue_order_strategy': 'priority',
+}