+++ src/startup_worker.sh
@@ -4,14 +4,14 @@ set -e
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 export PYTHONPATH=$SCRIPT_DIR
 
-# Set the ini file path 
+# Set the ini file path
 export OASIS_INI_PATH="${SCRIPT_DIR}/conf.ini"
 
 # Delete celeryd.pid file - fix que pickup issues on reboot of server
 rm -f /home/worker/celeryd.pid
 
 
-# ---- needs to be one or the other?? --- 
+# ---- needs to be one or the other?? ---
     # OLD - wait for it (test and remove)
     #./src/utils/wait-for-it.sh "$OASIS_RABBIT_HOST:$OASIS_RABBIT_PORT" -t 60
     # use for REDIS
@@ -23,4 +23,5 @@ rm -f /home/worker/celeryd.pid
 # celery --app src.model_execution_worker.tasks worker --concurrency=1 --loglevel=INFO -Q "${OASIS_MODEL_SUPPLIER_ID}-${OASIS_MODEL_ID}-${OASIS_MODEL_VERSION_ID}" |& tee -a /var/log/oasis/worker.log
 
 # Start new worker on init
-celery --app src.model_execution_worker.distributed_tasks worker --concurrency=1 --loglevel=INFO -Q "${OASIS_MODEL_SUPPLIER_ID}-${OASIS_MODEL_ID}-${OASIS_MODEL_VERSION_ID}" |& tee -a /var/log/oasis/worker.log
+celery --app src.model_execution_worker.distributed_tasks worker --loglevel=INFO -Q "${OASIS_MODEL_SUPPLIER_ID}-${OASIS_MODEL_ID}-${OASIS_MODEL_VERSION_ID}" |& tee -a /var/log/oasis/worker.log
+